+++ utils/defaultsettings.gd
@@ -13,7 +13,7 @@ var overlay_dimdownafter = 30
 var overlay_undimchime = true
 var overlay_chimedevice = "Default"
 var overlay_undimstare = true
-var overlay_undimstareseconds = 5
+var overlay_undimstareseconds = 3
 
 # Overlay position / rotation
 var overlay_size = 1.5